Implemented file caching for CSV and PDF downloads
Downloads are generated by a Sidekiq job and cached to disk
(public/downloads/). The files are named as a SHA1 digest of the job's
type and parameters: SHA1(PARAMS || TYPE). Future jobs with the same
params will check for the existence of a file with that hash, and return
that if said file exists.

This commit and prior has no implementation of cache sweeping, so until
next time, the downloads folder is gonna fill up like me at a free
buffet.