Fixed a bug with the Checklist that prevented cached downloads from being served.
When a cached download was found, Checklist.new() was not being called,
meaning that none of the global variables were being defined by
Checklist#initialize_params(). Checklist#summarise_filters relied on
these variables being available, and so Checklist#initialize_params() is
now called by itself when a cache is available.