fix: do not use setter where behaviour depends on multiple attributes
https://unep-wcmc.codebasehq.com/projects/species-rails-upgrade/tickets/117

`scientific_name=` was being called before `name_status` was set, which meant
that it always behaved as if `name_status` were the string 'A' during
instantiation, because the setter for `name_status` had not yet been called.

This meant that new synonyms were losing the genus part of the binomial
full name assigned to them, because this is stripped for A-names.